fix: queries and types

+++ src/main/java/com/example/clinic/dto/AppointmentTypeDTO.java
@@ -11,6 +11,6 @@ public record AppointmentTypeDTO (
     String name,
     String description,
     Integer duration,
-    Double price,
+    BigDecimal price,
     Doctor doctor
 ) {}
